fi.po 100%; additional special characters checks at read from config.cfg time (#1261)

+++ src/ct/ct_actions_edit.cc
@@ -320,7 +320,7 @@ void CtActions::_insert_toc_at_pos(Glib::RefPtr<Gtk::TextBuffer> text_buffer, co
         if (entry.is_node) {
             bullet_char = bullets_list[0];
         } else {
-            int bullet_index = entry.h_level + 1;
+            size_t bullet_index = entry.h_level + 1;
             if (bullet_index >= bullets_list.size()) {
                 bullet_index = bullets_list.size() - 1;
             }
